Read Person

+++ src/lesson7/PayEmployye/ContractEmployee.java
@@ -2,10 +2,10 @@ package lesson7.PayEmployye;
 
 public class ContractEmployee extends Employee implements Payment {
     private String federalTaxIdmember;
-    private double hourlyRate;
+    private int hourlyRate;
     private int numberOfHoursWorked;
 
-    public ContractEmployee(String employeeld, String federalTaxIdmember, double hourlyRate, int numberOfHoursWorked) {
+    public ContractEmployee(String employeeld, String federalTaxIdmember, int hourlyRate, int numberOfHoursWorked) {
         super(employeeld);
         this.federalTaxIdmember = federalTaxIdmember;
         this.hourlyRate = hourlyRate;
@@ -13,7 +13,7 @@ public class ContractEmployee extends Employee implements Payment {
     }
 
     @Override
-    public double calculatePay() {
+    public int calculatePay() {
         return hourlyRate * numberOfHoursWorked;
     }
