+++ src/lesson7/PayEmployye/Payment.java
@@ -1,6 +1,6 @@
 package lesson7.PayEmployye;
 
 public interface Payment {
-    public double calculatePay();
+    public int calculatePay();
 
 }
